# Improve QuerySelectorAll performance