# Types of Changes

## Prerequisites

Please make sure you can check the following two boxes:

- [x] I have read the **CONTRIBUTING** document
- [x] My code follows the code style of this project

## Contribution Type

What types of changes does your code introduce? Put an `x` in all the boxes that apply:

- [ ] Bug fix (non-breaking change which fixes an issue, please reference the issue id)
- [ ] New feature (non-breaking change which adds functionality, make sure to open an associated issue first)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] My change requires a change to the documentation
- [ ] I have updated the documentation accordingly
- [ ] I have added tests to cover my changes
- [x] All new and existing tests passed

## Description

#584 seemed like an interesting long-standing issue as optimizations are always fun. This PR basically reduces interface dispatch cost and optimizes node iteration for `MatchAll` and `MatchAny`.

* reduce memory usage in `NodeExtensions` by reusing stack instance, affects scenarios querying non-roots
* pass filter logic (and state) into `GetDescendantsAndSelf` to minimize items returned from enumerator
* force inline for `NodeList` members
* make `QueryExtensions` methods take `INodeList` input as generic type to remove virtual dispatch when call site knows the actual type
* update to latest BenchmarkDotNet, add `net6.0` target to benchmark project (realistic benchmarks for modern workloads)

fixes #584

## Benchmark results

```

BenchmarkDotNet v0.13.7, Windows 11 (10.0.23521.1000)
AMD Ryzen 9 5950X, 1 CPU, 32 logical and 16 physical cores
.NET SDK 8.0.100-preview.7.23376.3
  [Host]   : .NET 6.0.21 (6.0.2123.36311), X64 RyuJIT AVX2
  ShortRun : .NET 6.0.21 (6.0.2123.36311), X64 RyuJIT AVX2

Job=ShortRun  IterationCount=3  LaunchCount=1  
WarmupCount=3  

```

### AngleSharp.Benchmarks.SelectorBenchmark

| **Diff**|Selector|Mean|Error|Allocated|
|------- |-------|-------:|-------|-------:|
| Old |.note|162.2 μs|48.42 μs|17.04 KB|
| **New** | **.note** | **106.10 μs (-35%)** | **10.612 μs** | **17.04 KB (0%)** |
| Old |#title|159.3 μs|9.09 μs|16.8 KB|
| **New** | **#title** | **104.93 μs (-34%)** | **25.131 μs** | **16.8 KB (0%)** |
| Old |a[href]|160.1 μs|5.89 μs|21.05 KB|
| **New** | **a[href]** | **117.24 μs (-27%)** | **7.712 μs** | **21.05 KB (0%)** |
| Old |a[href][lang][class]|170.3 μs|10.04 μs|17.12 KB|
| **New** | **a[href][lang][class]** | **121.18 μs (-29%)** | **6.808 μs** | **17.12 KB (0%)** |
| Old |body|155.1 μs|5.56 μs|16.8 KB|
| **New** | **body** | **103.34 μs (-33%)** | **6.323 μs** | **16.8 KB (0%)** |
| Old |body div|167.9 μs|10.41 μs|24.5 KB|
| **New** | **body div** | **109.53 μs (-35%)** | **20.872 μs** | **24.5 KB (0%)** |
| Old |div|158.5 μs|2.75 μs|17.84 KB|
| **New** | **div** | **99.94 μs (-37%)** | **13.797 μs** | **17.84 KB (0%)** |
| Old |div #title|166.0 μs|31.90 μs|17.22 KB|
| **New** | **div #title** | **109.41 μs (-34%)** | **7.478 μs** | **17.22 KB (0%)** |
| Old |div + p|243.4 μs|6.31 μs|32.81 KB|
| **New** | **div + p** | **175.63 μs (-28%)** | **3.018 μs** | **32.8 KB (0%)** |
| Old |div > p|178.7 μs|11.65 μs|36.38 KB|
| **New** | **div > p** | **118.76 μs (-34%)** | **11.107 μs** | **36.38 KB (0%)** |
| Old |div > p > a|168.6 μs|8.40 μs|34.85 KB|
| **New** | **div > p > a** | **120.81 μs (-28%)** | **10.098 μs** | **34.85 KB (0%)** |
| Old |div ~ p|960.0 μs|156.94 μs|731.67 KB|
| **New** | **div ~ p** | **876.05 μs (-9%)** | **114.184 μs** | **731.67 KB (0%)** |
| Old |div p|189.9 μs|5.54 μs|62.18 KB|
| **New** | **div p** | **136.30 μs (-28%)** | **16.232 μs** | **62.18 KB (0%)** |
| Old |div p a|225.3 μs|18.13 μs|78.78 KB|
| **New** | **div p a** | **158.94 μs (-29%)** | **4.447 μs** | **78.78 KB (0%)** |
| Old |div, p, a|179.0 μs|3.70 μs|33.19 KB|
| **New** | **div, p, a** | **133.19 μs (-26%)** | **8.271 μs** | **33.19 KB (0%)** |
| Old |div:not(.example)|159.1 μs|1.91 μs|17.52 KB|
| **New** | **div:not(.example)** | **107.77 μs (-32%)** | **10.480 μs** | **17.52 KB (0%)** |
| Old |div.example|159.7 μs|82.68 μs|18.01 KB|
| **New** | **div.example** | **106.78 μs (-33%)** | **7.062 μs** | **18.01 KB (0%)** |
| Old |div.e(...).note [21]|179.1 μs|11.68 μs|18.39 KB|
| **New** | **div.e(...).note [21]** | **128.61 μs (-28%)** | **3.825 μs** | **18.39 KB (0%)** |
| Old |div[class!=made_up]|160.4 μs|24.60 μs|18.07 KB|
| **New** | **div[class!=made_up]** | **110.34 μs (-31%)** | **11.459 μs** | **18.07 KB (0%)** |
| Old |div[class]|153.6 μs|9.55 μs|18.02 KB|
| **New** | **div[class]** | **107.88 μs (-30%)** | **6.337 μs** | **18.02 KB (0%)** |
| Old |div[class*=e]|159.8 μs|13.64 μs|18.05 KB|
| **New** | **div[class*=e]** | **108.30 μs (-32%)** | **12.215 μs** | **18.05 KB (0%)** |
| Old |div[class^=exa]|170.7 μs|59.14 μs|18.06 KB|
| **New** | **div[class^=exa]** | **109.30 μs (-36%)** | **8.698 μs** | **18.06 KB (0%)** |
| Old |div[c(...)mple] [28]|167.5 μs|21.36 μs|18.18 KB|
| **New** | **div[c(...)mple] [28]** | **107.00 μs (-36%)** | **3.059 μs** | **18.18 KB (0%)** |
| Old |div[class=example]|162.4 μs|9.58 μs|18.09 KB|
| **New** | **div[class=example]** | **110.83 μs (-32%)** | **18.372 μs** | **18.09 KB (0%)** |
| Old |div[class\|=dialog]|153.9 μs|7.02 μs|16.98 KB|
| **New** | **div[class\|=dialog]** | **109.27 μs (-29%)** | **8.509 μs** | **16.98 KB (0%)** |
| Old |div[class~=example]|163.7 μs|6.93 μs|26.02 KB|
| **New** | **div[class~=example]** | **113.05 μs (-31%)** | **6.515 μs** | **26.02 KB (0%)** |
| Old |div[class$=mple]|157.7 μs|65.60 μs|18.06 KB|
| **New** | **div[class$=mple]** | **106.89 μs (-32%)** | **12.072 μs** | **18.06 KB (0%)** |
| Old |h1[id(...)tors) [26]|156.1 μs|11.43 μs|17.55 KB|
| **New** | **h1[id(...)tors) [26]** | **111.38 μs (-29%)** | **2.009 μs** | **17.59 KB (0%)** |
| Old |h1#title|168.8 μs|44.50 μs|16.97 KB|
| **New** | **h1#title** | **109.20 μs (-35%)** | **23.056 μs** | **16.97 KB (0%)** |
| Old |h1#title + div > p|230.8 μs|6.79 μs|38.77 KB|
| **New** | **h1#title + div > p** | **164.57 μs (-29%)** | **9.085 μs** | **38.77 KB (0%)** |
| Old |p:con(...)tors) [21]|360.0 μs|7.11 μs|234.51 KB|
| **New** | **p:con(...)tors) [21]** | **295.80 μs (-18%)** | **34.521 μs** | **247.17 KB (+5%)** |
| Old |p:first-child|163.6 μs|5.55 μs|17.98 KB|
| **New** | **p:first-child** | **115.14 μs (-30%)** | **17.388 μs** | **17.98 KB (0%)** |
| Old |p:last-child|165.7 μs|11.70 μs|17.46 KB|
| **New** | **p:last-child** | **112.19 μs (-32%)** | **2.951 μs** | **17.46 KB (0%)** |
| Old |p:nth-child(2n)|736.3 μs|58.25 μs|21.17 KB|
| **New** | **p:nth-child(2n)** | **674.85 μs (-8%)** | **80.690 μs** | **21.17 KB (0%)** |
| Old |p:nth-child(2n+1)|779.3 μs|90.54 μs|21.2 KB|
| **New** | **p:nth-child(2n+1)** | **696.89 μs (-11%)** | **116.719 μs** | **21.2 KB (0%)** |
| Old |p:nth-child(even)|745.9 μs|296.13 μs|21.15 KB|
| **New** | **p:nth-child(even)** | **678.07 μs (-9%)** | **13.369 μs** | **21.15 KB (0%)** |
| Old |p:nth-child(n)|791.3 μs|452.90 μs|25.17 KB|
| **New** | **p:nth-child(n)** | **676.86 μs (-14%)** | **11.371 μs** | **25.17 KB (0%)** |
| Old |p:nth-child(odd)|733.6 μs|60.70 μs|21.15 KB|
| **New** | **p:nth-child(odd)** | **673.66 μs (-8%)** | **15.086 μs** | **21.15 KB (0%)** |
| Old |p:only-child|285.9 μs|28.29 μs|16.96 KB|
| **New** | **p:only-child** | **222.05 μs (-22%)** | **49.198 μs** | **16.95 KB (0%)** |
| Old |ul .tocline2|193.7 μs|76.40 μs|18.84 KB|
| **New** | **ul .tocline2** | **108.13 μs (-44%)** | **4.266 μs** | **18.84 KB (0%)** |
| Old |ul.toc > li.tocline2|183.0 μs|48.48 μs|18.25 KB|
| **New** | **ul.toc > li.tocline2** | **113.93 μs (-38%)** | **17.205 μs** | **18.25 KB (0%)** |
| Old |ul.toc li.tocline2|178.3 μs|47.72 μs|19.16 KB|
| **New** | **ul.toc li.tocline2** | **117.20 μs (-34%)** | **29.391 μs** | **19.16 KB (0%)** |