@lahma 2023-08-20
I've returned the old methods which now call the generic versions. One method body I had to copy paste as compiler got too confused. Changes in the last commit which was amended.